Rework configuration as discussed on call.
